+++ packages/prism/lib/prism.dart
@@ -2,4 +2,5 @@ export 'src/ray_base.dart';
 export 'src/ray_rgb.dart';
 export 'src/ray_hsl.dart';
 export 'src/ray_oklab.dart';
+export 'src/ray_oklch.dart';
 export 'src/ray_scheme.dart';